make it nicer

+++ install.sh
@@ -50,11 +50,19 @@ if [[ -n "$install" ]]; then
     done
 fi
 
-rm -rf /tmp/dump978 &>/dev/null || true
-git clone --single-branch --depth 1 --branch master https://github.com/adsbxchange/uat2esnt.git /tmp/dump978
-cd /tmp/dump978/
-make uat2esnt
-cp uat2esnt $ipath
+UAT_REPO="https://github.com/adsbxchange/uat2esnt.git"
+UAT_VERSION="$(git ls-remote $UAT_REPO | grep HEAD | cut -f1)"
+if ! grep -e "$UAT_VERSION" -qs $ipath/uat2esnt_version; then
+    rm -rf /tmp/uat2esnt
+    git clone --single-branch --depth 1 --branch master $UAT_REPO /tmp/uat2esnt
+    cd /tmp/uat2esnt
+    make -j3 uat2esnt
+    rm -f $ipath/uat2esnt
+    cp uat2esnt $ipath
+    git rev-parse HEAD > $ipath/uat2esnt_version
+else
+    echo uat2esnt already current version
+fi
 
 READSB_REPO="https://github.com/adsbxchange/readsb.git"
 READSB_VERSION="$(git ls-remote $READSB_REPO | grep HEAD | cut -f1)"
@@ -63,10 +71,12 @@ if ! grep -e "$READSB_VERSION" -qs $ipath/readsb_version; then
     git clone --single-branch --depth 1 $READSB_REPO /tmp/readsb
     cd /tmp/readsb
     apt install -y libncurses5-dev zlib1g-dev zlib1g || true
-    make
+    make -j3 AIRCRAFT_HASH_BITS=12
     rm -f $ipath/readsb
     cp readsb $ipath
     git rev-parse HEAD > $ipath/readsb_version
+else
+    echo feed client already current version
 fi
 
 cd "$current_path"
